Introduce non-flat package namespace
I guess a flat namespace for everything is not going to work in the
long run, so let's try to introduce some sensible namespaces.

Also includes some check name and description tweaks for nicer and
more readable output.

Note: Can't use lib.string because string is already a namespace for
some rego builtins.