Introduce some unit tests
The tests are for just the rego under lib. I didn't add tests for
the checks themselves as yet.

So far, the unit tests are simple enough that I didn't feel the need
to bring in the `lib.assert_equal` and friends from ec-policies,
though I could could imagine it being added in the near future.